added global exception handling and validation on entity

+++ employee-management-system-backend/src/main/java/com/alok/employee/services/EmployeeServiceImpl.java
@@ -1,14 +1,18 @@
 package com.alok.employee.services;
 
+import com.alok.employee.customexception.EmployeeNotFoundException;
 import com.alok.employee.customexception.EmptyInputException;
+import com.alok.employee.customexception.EmptyListException;
 import com.alok.employee.model.Employee;
 import com.alok.employee.entity.EmployeeEntity;
 import com.alok.employee.repository.EmployeeRepository;
 import org.springframework.beans.BeanUtils;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -27,32 +31,43 @@ public class EmployeeServiceImpl implements EmployeeService{
             throw new EmptyInputException("666","input fields are null");
         }
         EmployeeEntity savedEmployee = employeeRepository.save(employee);
-//        savedEmployee=  employeeRepository.save(employee);
         return savedEmployee;
     }
 
     @Override
     public List<Employee> getAllEmployees() {
-        List<EmployeeEntity> employeeEntities
-                = employeeRepository.findAll();
-
-        List<Employee> employees = employeeEntities
-                .stream()
-                .map(emp -> new Employee(
-                        emp.getId(),
-                        emp.getFirstName(),
-                        emp.getLastName(),
-                        emp.getEmailId()))
-                .collect(Collectors.toList());
-        return employees;
+        List<EmployeeEntity> employeeEntities = employeeRepository.findAll();
+        try {
+            if (employeeEntities.isEmpty()){
+                throw new EmptyListException();
+            }
+            else {
+                List<Employee> employees = employeeEntities
+                        .stream()
+                        .map(emp -> new Employee(
+                                emp.getId(),
+                                emp.getFirstName(),
+                                emp.getLastName(),
+                                emp.getEmailId()))
+                        .collect(Collectors.toList());
+                return employees;
+            }
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
     }
 
     @Override
-    public boolean deleteEmployee(Long id) {
-        EmployeeEntity employee = employeeRepository.findById(id).get();
-        employeeRepository.delete(employee);
-        return true;
+    public boolean deleteEmployee(Long id) throws EmptyInputException {
+        EmployeeEntity employee = employeeRepository.findById(id).orElse(null);
+        if (employee != null) {
+            employeeRepository.delete(employee);
+            return true;
+        } else {
+            throw new EmptyInputException("555", "Employee not found for deletion");
+        }
     }
+
 
     @Override
     public EmployeeEntity getEmployeeById(Long id) {
@@ -66,13 +81,24 @@ public class EmployeeServiceImpl implements EmployeeService{
 
     @Override
     public Employee updateEmployee(Long id, Employee employee) {
-        EmployeeEntity employeeEntity
-                = employeeRepository.findById(id).get();
-        employeeEntity.setEmailId(employee.getEmailId());
-        employeeEntity.setFirstName(employee.getFirstName());
-        employeeEntity.setLastName(employee.getLastName());
 
-        employeeRepository.save(employeeEntity);
-        return employee;
+        try{
+            Optional<EmployeeEntity> optionalEmployeeEntity=employeeRepository.findById(id);
+            if(optionalEmployeeEntity.isPresent()){
+                EmployeeEntity employeeEntity=optionalEmployeeEntity.get();
+                employeeEntity.setEmailId(employee.getEmailId());
+                employeeEntity.setFirstName(employee.getFirstName());
+                employeeEntity.setLastName(employee.getLastName());
+                employeeRepository.save(employeeEntity);
+                return employee;
+            }
+            else {
+                throw new EmployeeNotFoundException();
+            }
+
+
+        } catch (Exception e) {
+            throw new RuntimeException("Error updating employee",e);
+        }
     }
 }
